feat: add a help model on the dashboard to explain the charts
This commit adds Zoa, an NHM under-development UI component lib to provide a modal. It is also now in use for the buttons. This commit also moves the a tag styling up to the main css and adds a no-link class for a tags which don't want this styling (such as the nav items and the dashboard title which links to GBIF).